+++ src/bin/cli.rs
@@ -3,7 +3,7 @@ use std::error::Error;
 use structopt::clap::ArgGroup;
 use structopt::StructOpt;
 
-use fints_institute_db::get_bank_by_bank_code;
+use fints_institute_db::{get_bank_by_bank_code, get_bank_by_bic};
 
 fn criterion_group() -> ArgGroup<'static> {
     ArgGroup::with_name("criterion").required(true)
@@ -19,12 +19,16 @@ fn criterion_group() -> ArgGroup<'static> {
 )]
 struct CliConfig {
     /// Look up bank by IBAN (format: DE02120300000000202051)
-    #[structopt(short, long, group = "criterion")]
+    #[structopt(long, group = "criterion")]
     iban: Option<Iban>,
 
     /// Look up bank by German bank code (format: 12030000)
-    #[structopt(short, long = "bankcode", group = "criterion")]
+    #[structopt(long = "bankcode", group = "criterion")]
     bank_code: Option<String>,
+
+    /// Look up bank by Bank Identifier Code (BIC) (format: GENODEM1MEN)
+    #[structopt(long = "bic", group = "criterion")]
+    bic: Option<String>,
 
     /// Change tool behavior to output all data for the record as JSON
     #[structopt(short = "j", long = "json")]
@@ -34,20 +38,21 @@ struct CliConfig {
 fn main() -> Result<(), Box<dyn Error>> {
     let args = CliConfig::from_args();
 
-    let bank_code = if let Some(iban) = args.iban {
+    let bank = if let Some(iban) = args.iban {
         if let Some(identifier) = iban.bank_identifier() {
-            identifier.to_string()
+            get_bank_by_bank_code(identifier)
         } else {
             eprintln!("Invalid or missing IBAN bank identifier");
             std::process::exit(1);
         }
+    } else if let Some(bank_code) = args.bank_code {
+        get_bank_by_bank_code(&bank_code)
+    } else if let Some(bic) = args.bic {
+        get_bank_by_bic(&bic)
     } else {
-        // It's safe to unwrap here because of the ArgParse group above which guarantees that one
-        // of these is set.
-        args.bank_code.unwrap()
+        unreachable!("Can't touch this.")
     };
 
-    let bank = get_bank_by_bank_code(&bank_code);
     if let Some(bank) = bank {
         if args.print_json {
             println!("{}", serde_json::to_string_pretty(&bank)?);