+++ tests/cli.rs
@@ -7,7 +7,7 @@ type Error = Box<dyn std::error::Error>;
 #[test]
 fn get_bank_by_bank_code() -> Result<(), Error> {
     Command::cargo_bin("cli")?
-        .arg("-b")
+        .arg("--bankcode")
         .arg("12030000")
         .assert()
         .success();
@@ -18,7 +18,7 @@ fn get_bank_by_bank_code() -> Result<(), Error> {
 #[test]
 fn invalid_bank_code() -> Result<(), Error> {
     Command::cargo_bin("cli")?
-        .arg("-b")
+        .arg("--bankcode")
         .arg("test")
         .assert()
         .stderr("No matching bank found\n")
@@ -30,7 +30,7 @@ fn invalid_bank_code() -> Result<(), Error> {
 #[test]
 fn get_bank_by_iban() -> Result<(), Error> {
     Command::cargo_bin("cli")?
-        .arg("-i")
+        .arg("--iban")
         .arg("DE02120300000000202051")
         .assert()
         .success();
@@ -41,10 +41,33 @@ fn get_bank_by_iban() -> Result<(), Error> {
 #[test]
 fn invalid_iban() -> Result<(), Error> {
     Command::cargo_bin("cli")?
-        .arg("-i")
+        .arg("--iban")
         .arg("test")
         .assert()
         .stderr("error: Invalid value for '--iban <iban>': the string does not follow the base IBAN rules\n")
+        .failure();
+
+    Ok(())
+}
+
+#[test]
+fn get_bank_by_bic() -> Result<(), Error> {
+    Command::cargo_bin("cli")?
+        .arg("--bic")
+        .arg("GENODEM1MEN")
+        .assert()
+        .success();
+
+    Ok(())
+}
+
+#[test]
+fn invalid_bic() -> Result<(), Error> {
+    Command::cargo_bin("cli")?
+        .arg("--bic")
+        .arg("test")
+        .assert()
+        .stderr("No matching bank found\n")
         .failure();
 
     Ok(())