+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+- Add `--bic` option to CLI tool.
 
 ## [1.1.0] - 2020-11-12
-- Add `get_bank_by_bic` to fetch a bank by its BIC
+- Add `get_bank_by_bic` to fetch a bank by its BIC.
 
 ## [1.0.6] - 2020-11-09
 - Actually publish all binaries.